Re-organized submission result view test cases and split them into different fixtures. Also moved some tests into separate files (such as output tests). Added some tests that were missing in the previous version. Updated student lookup tests to allow ultimate submission feedback instead of max for staff viewing students' ultimate submission. Still working on fixing up some of the moved tests.
