+++ autograder/rest_api/tests/test_views/test_submission_views/test_submission_result_views/test_submission_result_views.py
@@ -1,0 +1,836 @@
+import datetime
+import json
+
+from django.http import FileResponse, QueryDict
+from django.urls import reverse
+from django.utils import timezone
+from rest_framework import status
+from rest_framework.test import APIClient
+
+import autograder.core.models as ag_models
+import autograder.core.utils as core_ut
+import autograder.utils.testing.model_obj_builders as obj_build
+from autograder.core.models.get_ultimate_submissions import get_ultimate_submission
+from autograder.core.submission_feedback import update_denormalized_ag_test_results
+from autograder.core.tests.test_submission_feedback.fdbk_getter_shortcuts import (
+    get_cmd_fdbk, get_submission_fdbk)
+from autograder.utils.testing import UnitTestBase
+
+
+class _FeedbackTestsBase(UnitTestBase):
+    def setUp(self):
+        super().setUp()
+        self.client = APIClient()
+
+        self.ag_test_cmd = obj_build.make_full_ag_test_command(
+            set_arbitrary_points=True,
+            normal_fdbk_config={
+                'return_code_fdbk_level': ag_models.ValueFeedbackLevel.correct_or_incorrect,
+                'show_points': True,
+                'show_actual_stdout': True,
+            },
+            past_limit_submission_fdbk_config={
+                'stdout_fdbk_level': ag_models.ValueFeedbackLevel.correct_or_incorrect,
+                'show_points': True,
+                'show_actual_stderr': True,
+            },
+            ultimate_submission_fdbk_config={
+                'stdout_fdbk_level': ag_models.ValueFeedbackLevel.expected_and_actual,
+                'stderr_fdbk_level': ag_models.ValueFeedbackLevel.expected_and_actual
+            },
+            staff_viewer_fdbk_config={
+                'stderr_fdbk_level': ag_models.ValueFeedbackLevel.correct_or_incorrect,
+                'show_points': True,
+                'show_actual_stdout': True,
+            }
+        )
+        self.ag_test_case = self.ag_test_cmd.ag_test_case
+        self.ag_test_suite = self.ag_test_case.ag_test_suite
+        self.project = self.ag_test_suite.project
+        self.project.validate_and_update(
+            visible_to_students=True,
+            hide_ultimate_submission_fdbk=False,
+            closing_time=timezone.now() + datetime.timedelta(days=1)
+        )
+        self.course = self.project.course
+
+        self.student_group1 = obj_build.make_group(project=self.project)
+        self.student1 = self.student_group1.members.first()
+
+    def do_get_fdbk_test(self, client,
+                         submission: ag_models.Submission,
+                         fdbk_category: ag_models.FeedbackCategory):
+        query_params = QueryDict(mutable=True)
+        query_params.update({'feedback_category': fdbk_category.value})
+        url = (reverse('submission-results', kwargs={'pk': submission.pk})
+               + '?' + query_params.urlencode())
+        response = client.get(url)
+        self.assertEqual(status.HTTP_200_OK, response.status_code)
+        self.assertEqual(get_submission_fdbk(submission, fdbk_category).to_dict(),
+                         response.data)
+
+    def do_get_fdbk_permission_denied_test(self, client,
+                                           submission: ag_models.Submission,
+                                           fdbk_category: ag_models.FeedbackCategory):
+        query_params = QueryDict(mutable=True)
+        query_params.update({'feedback_category': fdbk_category.value})
+        url = (reverse('submission-results', kwargs={'pk': submission.pk})
+               + '?' + query_params.urlencode())
+        response = client.get(url)
+        self.assertEqual(status.HTTP_403_FORBIDDEN, response.status_code)
+
+    def do_get_output_and_diff_test(self, client,
+                                    submission: ag_models.Submission,
+                                    cmd_result: ag_models.AGTestCommandResult,
+                                    fdbk_category: ag_models.FeedbackCategory):
+        urls_and_field_names = self.get_output_and_diff_test_urls(
+            submission, cmd_result, fdbk_category)
+        for url, field_name in urls_and_field_names:
+            response = client.get(url)
+            self.assertEqual(status.HTTP_200_OK, response.status_code)
+            expected = getattr(get_cmd_fdbk(cmd_result, fdbk_category), field_name)
+            if isinstance(expected, core_ut.DiffResult):
+                actual = ('' if response.content == b''
+                          else json.loads(response.content.decode('utf-8')))
+                self.assertEqual(expected.diff_content, actual)
+            else:
+                if not isinstance(response, FileResponse):
+                    self.assertIsNone(response.data)
+                else:
+                    self.assertEqual(
+                        expected.read(), b''.join((chunk for chunk in response.streaming_content)))
+
+    def do_get_output_and_diff_permission_denied_test(self, client,
+                                                      submission: ag_models.Submission,
+                                                      cmd_result: ag_models.AGTestCommandResult,
+                                                      fdbk_category: ag_models.FeedbackCategory):
+        urls_and_field_names = self.get_output_and_diff_test_urls(
+            submission, cmd_result, fdbk_category)
+        for url, field_name in urls_and_field_names:
+            response = client.get(url)
+            self.assertEqual(status.HTTP_403_FORBIDDEN, response.status_code)
+
+    def get_output_and_diff_test_urls(self, submission: ag_models.Submission,
+                                      cmd_result: ag_models.AGTestCommandResult,
+                                      fdbk_category: ag_models.FeedbackCategory):
+        result = []
+        for field_name, url_lookup in self.OUTPUT_AND_DIFF_FIELDS_TO_URL_LOOKUPS.items():
+            query_params = QueryDict(mutable=True)
+            query_params.update({
+                'feedback_category': fdbk_category.value
+            })
+            url = (reverse(url_lookup,
+                           kwargs={'pk': submission.pk, 'result_pk': cmd_result.pk})
+                   + '?' + query_params.urlencode())
+            result.append((url, field_name))
+
+        return result
+
+    OUTPUT_AND_DIFF_FIELDS_TO_URL_LOOKUPS = {
+        'stdout': 'ag-test-cmd-result-stdout',
+        'stderr': 'ag-test-cmd-result-stderr',
+        'stdout_diff': 'ag-test-cmd-result-stdout-diff',
+        'stderr_diff': 'ag-test-cmd-result-stderr-diff',
+    }
+
+
+class NormalSubmissionFeedbackTestCase(_FeedbackTestsBase):
+    def setUp(self):
+        super().setUp()
+
+        self.student_group_normal_submission = obj_build.make_finished_submission(
+            group=self.student_group1)
+        self.normal_submission_result = obj_build.make_correct_ag_test_command_result(
+            self.ag_test_cmd, submission=self.student_group_normal_submission)
+        self.student_group_normal_submission = update_denormalized_ag_test_results(
+            self.student_group_normal_submission.pk)
+
+    def test_student_get_own_normal_submission_normal_fdbk(self):
+        self.client.force_authenticate(self.student1)
+        self.do_get_fdbk_test(self.client, self.student_group_normal_submission,
+                              ag_models.FeedbackCategory.normal)
+
+        self.do_get_output_and_diff_test(self.client, self.student_group_normal_submission,
+                                         self.normal_submission_result,
+                                         ag_models.FeedbackCategory.normal)
+
+    def test_student_get_own_normal_submission_non_normal_fdbk_permission_denied(self):
+        self.client.force_authenticate(self.student1)
+
+        # past_limit_submission
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            ag_models.FeedbackCategory.past_limit_submission)
+
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            self.normal_submission_result,
+            ag_models.FeedbackCategory.past_limit_submission)
+
+        # ultimate_submission
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            ag_models.FeedbackCategory.ultimate_submission)
+
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            self.normal_submission_result,
+            ag_models.FeedbackCategory.ultimate_submission)
+
+        # staff_viewer
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            ag_models.FeedbackCategory.staff_viewer)
+
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            self.normal_submission_result,
+            ag_models.FeedbackCategory.staff_viewer)
+
+        # max
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            ag_models.FeedbackCategory.max)
+
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            self.normal_submission_result,
+            ag_models.FeedbackCategory.max)
+
+    def test_student_get_other_student_submission_normal_fdbk_permission_denied(self):
+        student2 = obj_build.make_student_user(self.course)
+        self.client.force_authenticate(student2)
+
+        self.do_get_fdbk_permission_denied_test(self.client, self.student_group_normal_submission,
+                                                ag_models.FeedbackCategory.normal)
+
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            self.normal_submission_result, ag_models.FeedbackCategory.normal)
+
+    def test_student_get_normal_submission_max_or_staff_viewer_fdbk_permission_denied(self):
+        self.client.force_authenticate(self.student1)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            ag_models.FeedbackCategory.max)
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            self.normal_submission_result, ag_models.FeedbackCategory.max)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            ag_models.FeedbackCategory.staff_viewer)
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            self.normal_submission_result, ag_models.FeedbackCategory.staff_viewer)
+
+    def test_student_get_normal_submission_fdbk_project_not_visible_permission_denied(self):
+        self.project.validate_and_update(visible_to_students=False)
+        self.client.force_authenticate(self.student1)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            ag_models.FeedbackCategory.normal)
+
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            self.normal_submission_result, ag_models.FeedbackCategory.normal)
+
+    def test_staff_get_student_normal_submission_normal_fdbk_permission_denied(self):
+        staff = obj_build.make_staff_user(self.course)
+        self.client.force_authenticate(staff)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            ag_models.FeedbackCategory.normal)
+
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            self.normal_submission_result,
+            ag_models.FeedbackCategory.normal)
+
+    def test_handgrader_get_student_normal_submission_normal_fdbk_permission_denied(self):
+        handgrader = obj_build.make_handgrader_user(self.course)
+        self.client.force_authenticate(handgrader)
+
+        self.do_get_fdbk_permission_denied_test(self.client,
+                                                self.student_group_normal_submission,
+                                                ag_models.FeedbackCategory.normal)
+
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            self.normal_submission_result, ag_models.FeedbackCategory.normal)
+
+
+class PastLimitSubmissionFeedbackTestCase(_FeedbackTestsBase):
+    def setUp(self):
+        super().setUp()
+
+        self.student_group_past_limit_submission = obj_build.make_finished_submission(
+            group=self.student_group1, is_past_daily_limit=True)
+        self.past_limit_submission_result = obj_build.make_incorrect_ag_test_command_result(
+            self.ag_test_cmd, submission=self.student_group_past_limit_submission)
+
+    def test_student_get_own_past_limit_submission_past_limit_fdbk(self):
+        self.client.force_authenticate(self.student1)
+        self.do_get_fdbk_test(self.client, self.student_group_past_limit_submission,
+                              ag_models.FeedbackCategory.past_limit_submission)
+
+        self.do_get_output_and_diff_test(self.client, self.student_group_past_limit_submission,
+                                         self.past_limit_submission_result,
+                                         ag_models.FeedbackCategory.past_limit_submission)
+
+    def test_student_get_other_student_past_limit_submission_past_limit_fdbk_permissn_denied(self):
+        student2 = obj_build.make_student_user(self.course)
+        self.client.force_authenticate(student2)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group_past_limit_submission,
+            ag_models.FeedbackCategory.past_limit_submission)
+
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_past_limit_submission,
+            self.past_limit_submission_result,
+            ag_models.FeedbackCategory.past_limit_submission)
+
+    def test_student_get_own_past_limit_submission_normal_fdbk_permission_denied(self):
+        self.client.force_authenticate(self.student1)
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group_past_limit_submission,
+            ag_models.FeedbackCategory.normal)
+
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_past_limit_submission,
+            self.past_limit_submission_result,
+            ag_models.FeedbackCategory.normal)
+
+    def test_student_get_past_limit_submission_max_or_staff_viewer_fdbk_permission_denied(self):
+        self.client.force_authenticate(self.student1)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group_past_limit_submission,
+            ag_models.FeedbackCategory.max)
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_past_limit_submission,
+            self.past_limit_submission_result, ag_models.FeedbackCategory.max)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group_past_limit_submission,
+            ag_models.FeedbackCategory.staff_viewer)
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_past_limit_submission,
+            self.past_limit_submission_result, ag_models.FeedbackCategory.staff_viewer)
+
+    def test_staff_get_student_past_limit_submission_past_limit_fdbk_permission_denied(self):
+        staff = obj_build.make_staff_user(self.course)
+        self.client.force_authenticate(staff)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group_past_limit_submission,
+            ag_models.FeedbackCategory.past_limit_submission)
+
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_past_limit_submission,
+            self.past_limit_submission_result,
+            ag_models.FeedbackCategory.past_limit_submission)
+
+
+class UltimateSubmissionFeedbackTestCase(_FeedbackTestsBase):
+    def setUp(self):
+        super().setUp()
+
+        self.student_group1_best_submission = obj_build.make_finished_submission(
+            group=self.student_group1)
+        self.student1_best_res = obj_build.make_correct_ag_test_command_result(
+            self.ag_test_cmd, submission=self.student_group1_best_submission)
+        self.student_group1_best_submission = update_denormalized_ag_test_results(
+            self.student_group1_best_submission.pk)
+
+        self.student_group1_most_recent_submission = obj_build.make_finished_submission(
+            group=self.student_group1)
+        self.student1_most_recent_res = obj_build.make_incorrect_ag_test_command_result(
+            self.ag_test_cmd, submission=self.student_group1_most_recent_submission)
+        self.student_group1_most_recent_submission = update_denormalized_ag_test_results(
+            self.student_group1_most_recent_submission.pk)
+
+        self.project.validate_and_update(hide_ultimate_submission_fdbk=False, closing_time=None)
+
+    def test_student_get_own_ultimate_submission_ultimate_fdbk(self):
+        now = timezone.now()
+
+        self.project.validate_and_update(
+            ultimate_submission_policy=ag_models.UltimateSubmissionPolicy.most_recent)
+
+        self.client.force_authenticate(self.student1)
+
+        self.do_get_fdbk_test(self.client, self.student_group1_most_recent_submission,
+                              ag_models.FeedbackCategory.ultimate_submission)
+        self.do_get_output_and_diff_test(self.client, self.student_group1_most_recent_submission,
+                                         self.student1_most_recent_res,
+                                         ag_models.FeedbackCategory.ultimate_submission)
+
+        self.project.validate_and_update(
+            ultimate_submission_policy=ag_models.UltimateSubmissionPolicy.best)
+
+        self.do_get_fdbk_test(self.client, self.student_group1_best_submission,
+                              ag_models.FeedbackCategory.ultimate_submission)
+        self.do_get_output_and_diff_test(self.client, self.student_group1_best_submission,
+                                         self.student1_best_res,
+                                         ag_models.FeedbackCategory.ultimate_submission)
+
+        # Different combinations of the deadline being past and having an extension or not.
+        past_deadline = now - timezone.timedelta(2)
+        past_extension = now - timezone.timedelta(2)
+        for closing_time, extension in [(past_deadline, None),
+                                        (past_deadline, past_extension)]:
+            self.project.validate_and_update(closing_time=closing_time)
+            self.student_group1.validate_and_update(extended_due_date=extension)
+            self.do_get_fdbk_test(self.client, self.student_group1_best_submission,
+                                  ag_models.FeedbackCategory.ultimate_submission)
+            self.do_get_output_and_diff_test(self.client, self.student_group1_best_submission,
+                                             self.student1_best_res,
+                                             ag_models.FeedbackCategory.ultimate_submission)
+
+        # Late day finished
+        self.project.validate_and_update(closing_time=past_deadline)
+        self.student_group1.validate_and_update(extended_due_date=None)
+        self.student_group1.late_days_used = {self.student1.username: 1}
+        self.student_group1.save()
+
+        self.do_get_fdbk_test(self.client, self.student_group1_best_submission,
+                              ag_models.FeedbackCategory.ultimate_submission)
+        self.do_get_output_and_diff_test(self.client, self.student_group1_best_submission,
+                                         self.student1_best_res,
+                                         ag_models.FeedbackCategory.ultimate_submission)
+
+        # Extension and late day finished
+        self.project.validate_and_update(closing_time=past_deadline)
+        self.student_group1.validate_and_update(
+            extended_due_date=now - datetime.timedelta(hours=12))
+        self.student_group1.late_days_used = {self.student1.username: 1}
+        self.student_group1.save()
+
+        self.do_get_fdbk_test(self.client, self.student_group1_best_submission,
+                              ag_models.FeedbackCategory.ultimate_submission)
+        self.do_get_output_and_diff_test(self.client, self.student_group1_best_submission,
+                                         self.student1_best_res,
+                                         ag_models.FeedbackCategory.ultimate_submission)
+
+    def test_student_get_other_student_ultimate_submission_ultimate_fdbk_permission_denied(self):
+        self.assertEqual(ag_models.UltimateSubmissionPolicy.most_recent,
+                         self.project.ultimate_submission_policy)
+
+        student2 = obj_build.make_student_user(self.course)
+        self.client.force_authenticate(student2)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            ag_models.FeedbackCategory.ultimate_submission)
+
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            self.student1_most_recent_res, ag_models.FeedbackCategory.ultimate_submission)
+
+    def test_student_get_own_ultimate_submission_ultimate_fdbk_hidden_permission_denied(self):
+        self.assertEqual(ag_models.UltimateSubmissionPolicy.most_recent,
+                         self.project.ultimate_submission_policy)
+        self.assertIsNone(self.project.closing_time)
+
+        self.project.validate_and_update(hide_ultimate_submission_fdbk=True)
+        self.client.force_authenticate(self.student1)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            ag_models.FeedbackCategory.ultimate_submission)
+
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            self.student1_most_recent_res, ag_models.FeedbackCategory.ultimate_submission)
+
+    def test_student_get_ultimate_fdbk_but_deadline_not_past_permission_denied(self):
+        self.assertEqual(ag_models.UltimateSubmissionPolicy.most_recent,
+                         self.project.ultimate_submission_policy)
+        self.project.validate_and_update(
+            closing_time=timezone.now() + timezone.timedelta(minutes=3))
+
+        self.client.force_authenticate(self.student1)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            ag_models.FeedbackCategory.ultimate_submission)
+
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            self.student1_most_recent_res, ag_models.FeedbackCategory.ultimate_submission)
+
+    def test_student_get_ultimate_fdbk_but_extension_not_past_permission_denied(self):
+        self.assertEqual(ag_models.UltimateSubmissionPolicy.most_recent,
+                         self.project.ultimate_submission_policy)
+        self.project.validate_and_update(
+            closing_time=timezone.now() - timezone.timedelta(minutes=2))
+        self.student_group1.validate_and_update(
+            extended_due_date=timezone.now() + timezone.timedelta(minutes=3))
+
+        self.client.force_authenticate(self.student1)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            ag_models.FeedbackCategory.ultimate_submission)
+
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            self.student1_most_recent_res, ag_models.FeedbackCategory.ultimate_submission)
+
+    def test_student_get_ultimate_fdbk_but_has_unfinished_late_day_permission_denied(self):
+        self.assertEqual(ag_models.UltimateSubmissionPolicy.most_recent,
+                         self.project.ultimate_submission_policy)
+        self.project.validate_and_update(
+            closing_time=timezone.now() - timezone.timedelta(minutes=2))
+        self.student_group1.late_days_used = {self.student1.username: 1}
+        self.student_group1.save()
+
+        self.client.force_authenticate(self.student1)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            ag_models.FeedbackCategory.ultimate_submission)
+
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            self.student1_most_recent_res, ag_models.FeedbackCategory.ultimate_submission)
+
+        self.fail('code needs fixing')
+
+    def test_student_ultimate_submission_get_max_or_staff_viewer_fdbk_permission_denied(self):
+        self.assertEqual(ag_models.UltimateSubmissionPolicy.most_recent,
+                         self.project.ultimate_submission_policy)
+
+        self.client.force_authenticate(self.student1)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            ag_models.FeedbackCategory.max)
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            self.student1_most_recent_res, ag_models.FeedbackCategory.max)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            ag_models.FeedbackCategory.staff_viewer)
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            self.student1_most_recent_res, ag_models.FeedbackCategory.staff_viewer)
+
+    def test_staff_get_student_ultimate_submission_ultimate_fdbk(self):
+        self.assertEqual(ag_models.UltimateSubmissionPolicy.most_recent,
+                         self.project.ultimate_submission_policy)
+
+        staff = obj_build.make_staff_user(self.course)
+        self.client.force_authenticate(staff)
+
+        self.do_get_fdbk_test(self.client, self.student_group1_most_recent_submission,
+                              ag_models.FeedbackCategory.ultimate_submission)
+        self.do_get_output_and_diff_test(self.client, self.student_group1_most_recent_submission,
+                                         self.student1_most_recent_res,
+                                         ag_models.FeedbackCategory.ultimate_submission)
+
+    def test_staff_get_student_ultimate_submission_max_fdbk_permission_denied(self):
+        self.assertEqual(ag_models.UltimateSubmissionPolicy.most_recent,
+                         self.project.ultimate_submission_policy)
+
+        staff = obj_build.make_staff_user(self.course)
+        self.client.force_authenticate(staff)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            ag_models.FeedbackCategory.max)
+
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            self.student1_most_recent_res, ag_models.FeedbackCategory.max)
+
+
+class StaffOwnSubmissionFeedbackTestCase(_FeedbackTestsBase):
+    def test_staff_get_own_submission_any_fdbk(self):
+        staff_group = obj_build.make_group(
+            project=self.project, members_role=obj_build.UserRole.staff)
+        staff = staff_group.members.first()
+
+        staff_submission = obj_build.make_finished_submission(group=staff_group)
+        result = obj_build.make_correct_ag_test_command_result(
+            self.ag_test_cmd, submission=staff_submission)
+        staff_submission = update_denormalized_ag_test_results(staff_submission.pk)
+
+        self.client.force_authenticate(staff)
+
+        self.do_get_fdbk_test(self.client, staff_submission, ag_models.FeedbackCategory.normal)
+        self.do_get_output_and_diff_test(self.client, staff_submission, result,
+                                         ag_models.FeedbackCategory.normal)
+
+        self.do_get_fdbk_test(self.client, staff_submission,
+                              ag_models.FeedbackCategory.past_limit_submission)
+        self.do_get_output_and_diff_test(self.client, staff_submission, result,
+                                         ag_models.FeedbackCategory.past_limit_submission)
+
+        self.do_get_fdbk_test(self.client, staff_submission,
+                              ag_models.FeedbackCategory.staff_viewer)
+        self.do_get_output_and_diff_test(self.client, staff_submission, result,
+                                         ag_models.FeedbackCategory.staff_viewer)
+
+        self.do_get_fdbk_test(self.client, staff_submission,
+                              ag_models.FeedbackCategory.ultimate_submission)
+        self.do_get_output_and_diff_test(self.client, staff_submission, result,
+                                         ag_models.FeedbackCategory.ultimate_submission)
+
+        self.do_get_fdbk_test(self.client, staff_submission, ag_models.FeedbackCategory.max)
+        self.do_get_output_and_diff_test(self.client, staff_submission, result,
+                                         ag_models.FeedbackCategory.max)
+
+
+class StaffStudentLookupFeedbackTestCase(_FeedbackTestsBase):
+    def setUp(self):
+        super().setUp()
+
+        self.student_group_normal_submission = obj_build.make_finished_submission(
+            group=self.student_group1)
+        self.normal_submission_result = obj_build.make_correct_ag_test_command_result(
+            self.ag_test_cmd, submission=self.student_group_normal_submission)
+        self.student_group_normal_submission = update_denormalized_ag_test_results(
+            self.student_group_normal_submission.pk)
+
+        staff = obj_build.make_staff_user(self.course)
+        self.client.force_authenticate(staff)
+
+    def test_staff_get_student_submission_student_lookup_fdbk(self):
+        self.do_get_fdbk_test(self.client, self.student_group_normal_submission,
+                              ag_models.FeedbackCategory.staff_viewer)
+        self.do_get_output_and_diff_test(self.client, self.student_group_normal_submission,
+                                         self.normal_submission_result,
+                                         ag_models.FeedbackCategory.staff_viewer)
+
+    def test_staff_get_student_submission_non_staff_viewer_fdbk_permission_denied(self):
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            ag_models.FeedbackCategory.normal)
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            self.normal_submission_result, ag_models.FeedbackCategory.normal)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            ag_models.FeedbackCategory.past_limit_submission)
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            self.normal_submission_result, ag_models.FeedbackCategory.past_limit_submission)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            ag_models.FeedbackCategory.ultimate_submission)
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            self.normal_submission_result, ag_models.FeedbackCategory.ultimate_submission)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            ag_models.FeedbackCategory.max)
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group_normal_submission,
+            self.normal_submission_result, ag_models.FeedbackCategory.max)
+
+
+class StaffStudentLookupUltimateSubmissionFeedbackTestCase(_FeedbackTestsBase):
+    def setUp(self):
+        super().setUp()
+
+        staff = obj_build.make_staff_user(self.course)
+        self.client.force_authenticate(staff)
+
+        self.student_group1_best_submission = obj_build.make_finished_submission(
+            group=self.student_group1)
+        self.student1_best_res = obj_build.make_correct_ag_test_command_result(
+            self.ag_test_cmd, submission=self.student_group1_best_submission)
+        self.student_group1_best_submission = update_denormalized_ag_test_results(
+            self.student_group1_best_submission.pk)
+
+        self.student_group1_most_recent_submission = obj_build.make_finished_submission(
+            group=self.student_group1)
+        self.student1_most_recent_res = obj_build.make_incorrect_ag_test_command_result(
+            self.ag_test_cmd, submission=self.student_group1_most_recent_submission)
+        self.student_group1_most_recent_submission = update_denormalized_ag_test_results(
+            self.student_group1_most_recent_submission.pk)
+
+        self.project.validate_and_update(hide_ultimate_submission_fdbk=False, closing_time=None)
+
+    def test_staff_get_student_ultimate_submission_fdbk(self):
+        for closing_time in None, timezone.now() - timezone.timedelta(minutes=5):
+            self.project.validate_and_update(
+                closing_time=closing_time,
+                ultimate_submission_policy=ag_models.UltimateSubmissionPolicy.most_recent)
+
+            self.do_get_fdbk_test(self.client, self.student_group1_most_recent_submission,
+                                  ag_models.FeedbackCategory.ultimate_submission)
+            self.do_get_output_and_diff_test(
+                self.client, self.student_group1_most_recent_submission,
+                self.student1_most_recent_res, ag_models.FeedbackCategory.ultimate_submission)
+
+            self.project.validate_and_update(
+                ultimate_submission_policy=ag_models.UltimateSubmissionPolicy.best)
+
+            self.do_get_fdbk_test(self.client, self.student_group1_best_submission,
+                                  ag_models.FeedbackCategory.ultimate_submission)
+            self.do_get_output_and_diff_test(
+                self.client, self.student_group1_best_submission,
+                self.student1_best_res, ag_models.FeedbackCategory.ultimate_submission)
+
+    def test_staff_get_student_ultimate_submission_ultimate_fdbk_hidden_permission_denied(self):
+        self.assertEqual(ag_models.UltimateSubmissionPolicy.most_recent,
+                         self.project.ultimate_submission_policy)
+
+        self.project.validate_and_update(hide_ultimate_submission_fdbk=True)
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            ag_models.FeedbackCategory.ultimate_submission)
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            self.student1_most_recent_res, ag_models.FeedbackCategory.ultimate_submission)
+
+    def test_staff_get_student_ultimate_submission_fdbk_before_deadline_permission_denied(self):
+        self.assertEqual(ag_models.UltimateSubmissionPolicy.most_recent,
+                         self.project.ultimate_submission_policy)
+
+        self.project.validate_and_update(
+            closing_time=timezone.now() + timezone.timedelta(minutes=5))
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            ag_models.FeedbackCategory.ultimate_submission)
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            self.student1_most_recent_res, ag_models.FeedbackCategory.ultimate_submission)
+
+    def test_staff_get_student_ultimate_submission_fdbk_before_extension_permission_denied(self):
+        self.assertEqual(ag_models.UltimateSubmissionPolicy.most_recent,
+                         self.project.ultimate_submission_policy)
+
+        self.project.validate_and_update(
+            closing_time=timezone.now() - timezone.timedelta(minutes=2))
+        self.student_group1.validate_and_update(
+            extended_due_date=timezone.now() + timezone.timedelta(minutes=2))
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            ag_models.FeedbackCategory.ultimate_submission)
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            self.student1_most_recent_res, ag_models.FeedbackCategory.ultimate_submission)
+
+    def test_staff_get_student_ultimate_submision_fdbk_pending_late_day_permission_denied(self):
+        self.assertEqual(ag_models.UltimateSubmissionPolicy.most_recent,
+                         self.project.ultimate_submission_policy)
+
+        self.project.validate_and_update(
+            closing_time=timezone.now() - timezone.timedelta(minutes=2))
+        self.assertIsNone(self.student_group1.extended_due_date)
+
+        self.student_group1.late_days_used = {self.student1.username: 1}
+        self.student_group1.save()
+
+        self.do_get_fdbk_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            ag_models.FeedbackCategory.ultimate_submission)
+        self.do_get_output_and_diff_permission_denied_test(
+            self.client, self.student_group1_most_recent_submission,
+            self.student1_most_recent_res, ag_models.FeedbackCategory.ultimate_submission)
+
+
+class AdminLookupStudentSubmissionTestCase(UnitTestBase):
+    def test_admin_get_student_submission_any_fdbk(self):
+        self.fail()
+
+
+class MiscSubmissionFeedbackTestCase(UnitTestBase):
+    def setUp(self):
+        super().setUp()
+
+        project = obj_build.make_project()
+        self.course = project.course
+
+        staff_group = obj_build.make_group(
+            project=project, members_role=obj_build.UserRole.staff)
+        self.staff = staff_group.members.first()
+        self.staff_normal_submission = obj_build.make_finished_submission(group=staff_group)
+
+    def test_invalid_fdbk_category_requested(self):
+        self.client.force_authenticate(self.staff)
+        query_params = QueryDict(mutable=True)
+        query_params.update({'feedback_category': 'not a value'})
+        url = (reverse('submission-results', kwargs={'pk': self.staff_normal_submission.pk})
+               + '?' + query_params.urlencode())
+        response = self.client.get(url)
+        self.assertEqual(status.HTTP_400_BAD_REQUEST, response.status_code)
+        self.assertIn('feedback_category', response.data)
+
+    def test_missing_fdbk_category(self):
+        self.client.force_authenticate(self.staff)
+        url = reverse('submission-results',
+                      kwargs={'pk': self.staff_normal_submission.pk})
+        response = self.client.get(url)
+        self.assertEqual(status.HTTP_400_BAD_REQUEST, response.status_code)
+        self.assertIn('feedback_category', response.data)
+
+
+class UltimateSubmissionWithLateDaysTestCase(UnitTestBase):
+    def test_group_member_with_different_ultimate_submission(self):
+        """
+        When a group member has a different ultimate submission than the
+        rest of the group due to running out of late days,
+        we need to make sure that the student can request ultimate
+        submission feedback on their ultimate submission.
+        """
+        project = obj_build.make_project(
+            visible_to_students=True, hide_ultimate_submission_fdbk=False,
+            ultimate_submission_policy=ag_models.UltimateSubmissionPolicy.most_recent)
+        group = obj_build.make_group(num_members=2, project=project)
+
+        doesnt_count_for_user, counts_for_user = group.members.all()
+
+        first_submission = obj_build.make_finished_submission(group=group)
+        most_recent_submission = obj_build.make_finished_submission(
+            group=group, does_not_count_for=[doesnt_count_for_user.username])
+
+        self.assertEqual(first_submission,
+                         get_ultimate_submission(group, user=doesnt_count_for_user))
+        self.assertEqual(most_recent_submission,
+                         get_ultimate_submission(group, user=counts_for_user))
+
+        client = APIClient()
+
+        client.force_authenticate(counts_for_user)
+        response = client.get(
+            self._make_submission_result_url(most_recent_submission,
+                                        ag_models.FeedbackCategory.ultimate_submission))
+        self.assertEqual(status.HTTP_200_OK, response.status_code)
+
+        client.force_authenticate(doesnt_count_for_user)
+        # Shouldn't be able to get ultimate submission fdbk on
+        # the most recent submission (that doesn't count for this user)
+        response = client.get(
+            self._make_submission_result_url(most_recent_submission,
+                                        ag_models.FeedbackCategory.ultimate_submission))
+        self.assertEqual(status.HTTP_403_FORBIDDEN, response.status_code)
+
+        response = client.get(
+            self._make_submission_result_url(first_submission,
+                                        ag_models.FeedbackCategory.ultimate_submission))
+        self.assertEqual(status.HTTP_200_OK, response.status_code)
+
+    def _make_submission_result_url(
+            self,
+            submission: ag_models.Submission,
+            fdbk_category: ag_models.FeedbackCategory=ag_models.FeedbackCategory.normal):
+        url = reverse('submission-results', kwargs={'pk': submission.pk})
+        url += '?feedback_category={}'.format(fdbk_category.value)
+
+        return url
